+++ src/main/kotlin/hilfling/backend/hilfling/controller/PositionController.kt
@@ -1,11 +1,7 @@
 package hilfling.backend.hilfling.controller
 
-import hilfling.backend.hilfling.model.Album
 import hilfling.backend.hilfling.model.Position
-import hilfling.backend.hilfling.repository.AlbumRepository
 import hilfling.backend.hilfling.repository.PositionRepository
-import me.liuwj.ktorm.database.Database
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.*
 
 @RestController
@@ -14,7 +10,7 @@ class PositionController {
     val repository = PositionRepository()
 
     @GetMapping("/{id}")
-    fun getById(@PathVariable("id") id: Long) : Position? {
+    fun getById(@PathVariable("id") id: Int) : Position? {
         return repository.findById(id)
     }
